requirements.txt
flask==3.1.1
selenium==4.34.2
beautifulsoup4==4.13.4
requests==2.32.4
python-dotenv==1.0.1
